use client was removed from the cart page and added to the component, and the cartpage and swiper were rearranged.
